+++ src/app/layout.tsx
@@ -4,8 +4,8 @@ import '@/styles/globals.css'
 import { Background } from '@/components/layout/background'
 import { Footer } from '@/components/layout/footer'
 import { Header } from '@/components/layout/header'
-import { buildThemeScript } from '@/lib/theme-runtime'
 import { buildPresenceScript } from '@/utils/achievements'
+import { buildThemeScript } from '@/utils/theme-runtime'
 import { type Metadata } from 'next'
 import { Noto_Sans, Space_Grotesk } from 'next/font/google'
