+++ src/components/layout/about/aboutme/mode-toggle-link.tsx
@@ -1,14 +1,13 @@
 'use client'
 
 import { useTheme } from '@/components/providers/theme-provider'
-import { useCallback, useMemo } from 'react'
-
 import { Button } from '@/components/ui/button'
 import { useThemeTransition } from '@/components/ui/theme-toggle'
 import { captureDarkModeEasterEgg } from '@/hooks/posthog'
-import { buildPerThemeVariantCss } from '@/lib/theme-css'
 import { themes } from '@/lib/themes'
-import { isSafari } from '@/lib/utils'
+import { buildPerThemeVariantCss } from '@/utils/theme-css'
+import { isSafari } from '@/utils/utils'
+import { useCallback, useMemo } from 'react'
 
 export function ModeToggleNote() {
   const noteCss = useMemo(() => {
