+++ src/components/layout/header/mobile-nav.tsx
@@ -5,7 +5,7 @@ import { useAchievements } from '@/components/providers/achievements-provider'
 import { Button } from '@/components/ui/button'
 import { useThemeTransition } from '@/components/ui/theme-toggle'
 import { NAVIGATION } from '@/lib/navmenu'
-import { cn } from '@/lib/utils'
+import { cn } from '@/utils/utils'
 import { MenuIcon, PawPrint, Trophy } from 'lucide-react'
 import type { Route } from 'next'
 import { usePathname, useRouter } from 'next/navigation'
