+++ src/components/layout/home/hero/profile-image.tsx
@@ -8,9 +8,9 @@ import Grumpy from '@/images/headshot/cartoon-grumpy.webp'
 import Ladybird from '@/images/headshot/cartoon-ladybird.webp'
 import { type AchievementId } from '@/lib/achievements'
 import { HOME_CONTENT } from '@/lib/content'
-import { buildPerThemeVariantCss } from '@/lib/theme-css'
 import { themes } from '@/lib/themes'
 import { isLadybirdUA } from '@/utils/ladybird'
+import { buildPerThemeVariantCss } from '@/utils/theme-css'
 import { getThemeHeadshot } from '@/utils/themes'
 import Image, { type StaticImageData } from 'next/image'
 import { useCallback, useEffect, useMemo, useState, type KeyboardEvent } from 'react'
