+++ src/components/ui/link-button.tsx
@@ -1,5 +1,5 @@
 import { Button, type buttonVariants } from '@/components/ui/button'
-import { cn } from '@/lib/utils'
+import { cn } from '@/utils/utils'
 import type { VariantProps } from 'class-variance-authority'
 import type { Route } from 'next'
 import Link from 'next/link'
