+++ src/components/ui/navigation-menu.tsx
@@ -3,7 +3,7 @@ import { cva } from 'class-variance-authority'
 import { ChevronDownIcon } from 'lucide-react'
 import * as React from 'react'
 
-import { cn } from '@/lib/utils'
+import { cn } from '@/utils/utils'
 
 function NavigationMenu({
   className,
