+++ src/components/ui/section-label.tsx
@@ -1,4 +1,4 @@
-import { cn } from '@/lib/utils'
+import { cn } from '@/utils/utils'
 
 type SectionLabelProps = {
   children: string
