+++ src/components/ui/theme-toggle.tsx
@@ -4,11 +4,11 @@ import { useTheme } from '@/components/providers/theme-provider'
 import { Button } from '@/components/ui/button'
 import { Tooltip, TooltipContent, TooltipTrigger } from '@/components/ui/tooltip'
 import { captureThemeChanged } from '@/hooks/posthog'
-import { buildPerThemeVariantCss } from '@/lib/theme-css'
-import { getAvailableThemes, getDefaultThemeForNow } from '@/lib/theme-runtime'
 import { themes, type Theme } from '@/lib/themes'
-import { cn, isSafari } from '@/lib/utils'
+import { buildPerThemeVariantCss } from '@/utils/theme-css'
+import { getAvailableThemes, getDefaultThemeForNow } from '@/utils/theme-runtime'
 import { getThemeIcon, getThemeLabel } from '@/utils/themes'
+import { cn, isSafari } from '@/utils/utils'
 import { Laptop, Smartphone } from 'lucide-react'
 import type { KeyboardEvent as ReactKeyboardEvent, MouseEvent as ReactMouseEvent } from 'react'
 import { useCallback, useEffect, useMemo, useRef, useState, type ComponentType } from 'react'
