+++ src/types/presence-bundle.d.ts
@@ -1,3 +1,3 @@
-declare module '@/lib/presence-bundle' {
+declare module '@/utils/presence-bundle' {
   export const PRESENCE_RUNTIME_BUNDLE: string
 }
