+++ src/types/theme-bundle.d.ts
@@ -1,3 +1,3 @@
-declare module '@/lib/theme-bundle' {
+declare module '@/utils/theme-bundle' {
   export const THEME_RUNTIME_BUNDLE: string
 }
