+++ src/utils/presence-script.ts
